Homework PDF_normal_distribution finished
