+++ 04-Makefiles/Computing the pdf for the normal distribution/automatizar.sh
@@ -1,0 +1,4 @@
+make run_seed_1
+make run_seed_2
+make run_seed_5
+make %.pdf
